requirements.txt
streamlit>=1.35.0
streamlit-webrtc>=0.47.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.26.0
matplotlib>=3.8.0
pandas>=2.1.0
av>=11.0.0
